da vedere bug

+++ src/main/java/BuildWeekEpicEnergyServices/entities/Cliente.java
@@ -36,7 +36,7 @@ public class Cliente {
     private LocalDate ultimoContattoIl;
 
     @Column(name = "fatturato_annuo")
-    private double fatturatoAnnuo;
+    private int fatturatoAnnuo;
 
     private String pec;
